@@ -1,6 +1,6 @@
 import React from 'react';
 import classnames from 'classnames';
-import { Scrollbars } from 'react-custom-scrollbars';
+import { Scrollbars } from 'react-custom-scrollbars-2';
 import SearchList from '../search-list';
 import NodeListGroups from './node-list-groups';
 import NodeListTree from './node-list-tree';
